Add new file for Decorators classes
