+++ person.rb
@@ -1,4 +1,5 @@
 require './nameable'
+require './decorators'
 
 class Person < Nameable
   def initialize(age, name = 'Unknown', parent_permission: true)
@@ -24,30 +25,6 @@ class Person < Nameable
 
   def of_age?
     @age >= 18
-  end
-end
-
-class Decorator < Nameable
-  def initialize(nameable)
-    super()
-    @nameable = nameable
-  end
-
-  def correct_name
-    @nameable.correct_name
-  end
-end
-
-class CapitalizeDecorator < Decorator
-  def correct_name
-    @nameable.correct_name.capitalize
-  end
-end
-
-class TrimmerDecorator < Decorator
-  def correct_name
-    trimmed_name = @nameable.correct_name[0..9]
-    trimmed_name.length == 10 ? trimmed_name : @nameable.correct_name
   end
 end
